+++ src/idv/jlchntoz/lang/Language.java
@@ -1,23 +1,23 @@
-package idv.jlchntoz;
+package idv.jlchntoz.lang;
 
 import java.util.Locale;
 
-public class Lang {
+public class Language {
 	
-	public static LangParent refresh(){
+	public static DefaultLang getLanguage(){
 		Locale locale = Locale.getDefault();  
 		String l=locale.getLanguage().toLowerCase();  
 		String c=locale.getCountry().toLowerCase();  
 		if(l.equalsIgnoreCase("zh")){
 			if(c.equalsIgnoreCase("cn")){
-				//zh_cn();
-				return new LangChineseSimp();
+				//zh_cn
+				return new ChineseSimpLang();
 			}
-			//zh();
-			return new LangChineseTran();
+			//zh
+			return new ChineseTranLang();
 		}
-		//en();
-		return new LangParent();
+		//en
+		return new DefaultLang();
 	}
